Adding Car Navigation Project to the created Repo
This project is the unfinished version from class on November 18th,
2014. Today we added in key functions that rotate and translate the car
as well as camera. The next step in this project is to add other
functions to our camera. We already have dolly (zoom in/zoom out
translation) and track (left/right translation) of the camera. We will
be adding tumble soon...